backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=0.27.0
numpy>=1.24.0
faiss-cpu>=1.7.3
langchain>=0.3.1
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.1
openai>=1.6.0
streamlit>=1.35.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.8.0
email-validator>=2.0.0
pydantic-settings>=2.0.0